+++ app/src/main/java/br/com/bvilela/listbuilder/service/notification/impl/NotifyAudienceServiceImpl.java
@@ -1,10 +1,13 @@
-package br.com.bvilela.listbuilder.service.notification;
+package br.com.bvilela.listbuilder.service.notification.impl;
 
 import br.com.bvilela.lib.model.CalendarEvent;
 import br.com.bvilela.listbuilder.config.NotifyProperties;
 import br.com.bvilela.listbuilder.enuns.ListTypeEnum;
 import java.time.LocalDate;
 import java.util.List;
+
+import br.com.bvilela.listbuilder.service.notification.NotifyAudienceService;
+import br.com.bvilela.listbuilder.service.notification.NotifyUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
