+++ app/src/test/java/br/com/bvilela/listbuilder/service/notification/SendNotificationServiceImplTest.java
@@ -7,6 +7,11 @@ import br.com.bvilela.listbuilder.builder.designacao.DesignacaoWriterDtoBuilder;
 import br.com.bvilela.listbuilder.dto.designacao.writer.DesignacaoWriterDTO;
 import br.com.bvilela.listbuilder.dto.limpeza.FinalListLimpezaDTO;
 import br.com.bvilela.listbuilder.dto.limpeza.FinalListLimpezaItemLayout2DTO;
+import br.com.bvilela.listbuilder.service.notification.impl.NotifyAudienceServiceImpl;
+import br.com.bvilela.listbuilder.service.notification.impl.NotifyChristianLifeServiceImpl;
+import br.com.bvilela.listbuilder.service.notification.impl.NotifyClearingServiceImpl;
+import br.com.bvilela.listbuilder.service.notification.impl.NotifyDesignationServiceImpl;
+import br.com.bvilela.listbuilder.service.notification.impl.SendNotificationServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
